CRUD feature for testimony form

+++ controllers/testimonyFormController.js
@@ -44,38 +44,114 @@ export const getAllTestimonyForm = async (req, res) => {
 }
 
 export const getTestimonyForm = async (req, res) => {
-  const testimonyForm = await prisma.testimonyForm.findById(req.params.id)
-  res.status(StatusCodes.OK).json({ testimonyForm })
+  try {
+    const testimonyForm = await prisma.testimonyForm.findUnique({
+      where: { id: parseInt(req.params.id) },
+    })
+
+    if (!testimonyForm) {
+      return res
+        .status(StatusCodes.NOT_FOUND)
+        .json({ error: 'Testimony form not found' })
+    }
+
+    res.status(StatusCodes.OK).json({ testimonyForm })
+  } catch (error) {
+    console.error('Error:', error)
+    return res
+      .status(StatusCodes.INTERNAL_SERVER_ERROR)
+      .json({ error: 'Internal Server Error' })
+  }
 }
 
 export const updateTestimonyForm = async (req, res) => {
   try {
-    const updatedTestimonyForm = await prisma.testimonyForm.update({
-      where: { id: parseInt(req.params.id) },
-      data: req.body,
+    const { id } = req.params
+    const parsedId = parseInt(id)
+
+    if (isNaN(parsedId)) {
+      return res.status(StatusCodes.BAD_REQUEST).json({ error: 'Invalid ID' })
+    }
+
+    // console.log('parsedId:', parsedId)
+
+    const existingTestimony = await prisma.testimonyForm.findUnique({
+      where: { id: parsedId },
     })
+
+    // console.log('existingTestimony:', existingTestimony)
+
+    if (!existingTestimony) {
+      return res
+        .status(StatusCodes.NOT_FOUND)
+        .json({ error: 'Testimony not found' })
+    }
+
+    const updatedData = { ...req.body }
+
+    // console.log('updatedData:', updatedData)
+
+    // console.log('req file:', req.file)
+
+    // Check if there's a file uploaded
+    if (req.file) {
+      const imageInfo = formatImage(req.file)
+      const cloudinaryResponse = await cloudinary.v2.uploader.upload(imageInfo)
+      updatedData.avatar = cloudinaryResponse.secure_url
+      updatedData.avatarPublicId = cloudinaryResponse.public_id
+    }
+
+    const updatedTestimony = await prisma.testimonyForm.update({
+      where: { id: parsedId },
+      data: updatedData,
+    })
+
+    console.log('updated Testimony:', updatedTestimony)
+
+    // Delete previous image from Cloudinary if a new one was uploaded
+    if (req.file && existingTestimony.avatarPublicId) {
+      await cloudinary.v2.uploader.destroy(existingTestimony.avatarPublicId)
+    }
 
     res
       .status(StatusCodes.OK)
-      .json({ msg: 'Testimony modified', updatedTestimonyForm })
+      .json({ msg: 'Testimony modified', updatedTestimony })
   } catch (error) {
     console.error('Error:', error)
-    return res.status(500).json({ error: 'Internal Server Error' })
+    return res
+      .status(StatusCodes.INTERNAL_SERVER_ERROR)
+      .json({ error: 'Internal Server Error' })
   }
 }
 
 
 export const deleteTestimonyForm = async (req, res) => {
   try {
-    const removedTestimony = await prisma.testimonyForm.delete({
+    const deletedTestimony = await prisma.testimonyForm.findUnique({
       where: { id: parseInt(req.params.id) },
     })
 
+    if (!deletedTestimony) {
+      return res
+        .status(StatusCodes.NOT_FOUND)
+        .json({ error: 'Testimony not found' })
+    }
+
+    await prisma.testimonyForm.delete({
+      where: { id: parseInt(req.params.id) },
+    })
+
+    if (deletedTestimony.avatarPublicId) {
+      await cloudinary.v2.uploader.destroy(deletedTestimony.avatarPublicId)
+    }
+
     res
       .status(StatusCodes.OK)
-      .json({ msg: 'Testimony deleted', testimony: removedTestimony })
+      .json({ msg: 'Testimony deleted', testimony: deletedTestimony })
   } catch (error) {
     console.error('Error:', error)
-    return res.status(500).json({ error: 'Internal Server Error' })
+    return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
+      error: 'Internal Server Error',
+    })
   }
 }
